Fixed mongoose methods for projs, added usrsRouter

+++ src/repository/mongoose.js
@@ -3,6 +3,7 @@ const debug = require('debug')('app:mongoose');
 
 const GitRepo = require('../models/gitRepo');
 const Project = require('../models/project');
+const user = require('../models/user');
 const User = require('../models/user');
 
 // connection
@@ -15,21 +16,51 @@ mongoose.connect('mongodb://localhost:27017/Github_Timelines_App_2_DB').then(() 
 
 
 
+
+
+
+
+// TIE STO SE TESTIRANI VEKJE IMAAT *** NAD NIV 
+
+// TODO: SEGA ZA SEGA NE CUVAS I COMMITS ZA REPO, MOZEBI KE E PREGOLEMO ZA VO DB
+
+
+
+
+
 // methods
 
 // -- Projects
-const createNewProject = async (req) => {       // TODO: treba da stava creator = username od session zemeno
+
+// ***
+const createNewProject = async (req) => {  
+    
+    // const creator = req.session.userName;
+    const creator = 'NekojCovek';               // FIXME: radi postman
     
     const newProject = new Project({
         name: req.body.name,
         status: req.body.status,
-        segments: req.body.segments
+        segments: req.body.segments,
+        creator           
     });
 
-    debug(newProject);
+    // debug(newProject);      
+
 
     try {
         const result = await newProject.save();
+
+        // Go stava proektot i kaj userot vo db
+        const theUser = await User.findOne({ gitUsername: creator });
+        theUser.projectIds.push(result.id);
+
+        const userObjectId = mongoose.Types.ObjectId(theUser.id);
+
+        // eslint-disable-next-line no-unused-vars
+        const resUsr = await User.findByIdAndUpdate(userObjectId, theUser, { new: true }).exec();
+        // debug(resUsr);
+
         return result;
     } catch (error) {
         debug(error);
@@ -37,11 +68,12 @@ const createNewProject = async (req) => {       // TODO: treba da stava creator 
     }
     
 };
-
+// ***
 const editProject = async (req) => {
     const { id } = req.params;
 
     const newProject = {            // NOTE: ne pravis new Project() kako pri addNewProject
+        creator: req.body.creator,
         name: req.body.name,
         status: req.body.status,
         segments: req.body.segments
@@ -57,12 +89,12 @@ const editProject = async (req) => {
         return `Cannot edit the project with id ${id}`;
     }
 };
-
+// ***          // preku ova dodavas/trgas segments, a za repos imas posebni metodi
 const getAllProjects = async () => {
     const result = await Project.find().exec();
     return result;
 };
-
+// ***
 const getProjectById = async (req) => {
     const { id } = req.params;
     try {
@@ -73,9 +105,47 @@ const getProjectById = async (req) => {
         return `Cannot find the project with id ${id}`;
     }
 };
+// ***
+const getProjectsOfLoggedInUser = async () => {
+    
+    // const username = req.session.userName;       // FIXME: Radi postman vaka
+    const username = 'NekojCovek';
 
-const deleteProject = async (req) => {
+    const allProjs = await Project.find().exec();
+
+    const result = allProjs.filter(pr => {
+        return pr.creator === username;
+    });
+
+    return result;
+};
+// ***
+const deleteProject = async (req) => {         // Userot moze da napravi delete samo na proekt na koj toj e kreator
     const { id } = req.params;
+
+    // const username = req.session.userName;       // FIXME: Radi postman vaka
+    const username = 'NekojCovek';
+    const theUser = await User.findOne({ gitUsername: username });
+
+
+    if (!theUser.projectIds.includes(id)) {
+        return 'Cannot delete a project that is not created by the logged in user';
+    }
+
+    try {       // delete project id from the user
+        const idxOfProjId = theUser.projectIds.indexOf(id);
+        theUser.projectIds.splice(idxOfProjId, 1);
+
+        const userObjectId = mongoose.Types.ObjectId(theUser.id);
+
+        const result = await User.findByIdAndUpdate(userObjectId, theUser, { new: true }).exec();
+        debug(result);
+
+    } catch (error) {
+        debug(error);
+        return 'Cannot remove project form user';
+    }
+
     try {
         const result = await Project.findByIdAndDelete(id).exec();
         return result;
@@ -87,6 +157,7 @@ const deleteProject = async (req) => {
 };
  
 // -- Users
+// ***
 const createNewUser = async (gitUsername, gitUrl, gitAvatarUrl, projectIds, followingGitRepos) => {
 
     const newUser = new User({
@@ -105,6 +176,19 @@ const createNewUser = async (gitUsername, gitUrl, gitAvatarUrl, projectIds, foll
         return error;
     }
 };
+// ***
+const getUserById = async (req) => {
+    const { id } = req.params;
+    try {
+        const res = await User.findById(id).exec();
+        return res;
+    } catch (error) {
+        debug(error);
+        return `Cannot find user with id: ${id}`;
+    }
+    
+};
+
 
 
 // -- Git Repos
@@ -183,6 +267,8 @@ const removeRepoFromProject = async (req) => {
     }
 };
 
+const addNoteToRepo = async (req) => {};
+
 
 
 
@@ -233,8 +319,10 @@ module.exports.getAllProjects = getAllProjects;
 module.exports.getProjectById = getProjectById;
 module.exports.deleteProject = deleteProject;
 module.exports.editProject = editProject;
+module.exports.getProjectsOfLoggedInUser = getProjectsOfLoggedInUser;
 
 module.exports.createNewUser = createNewUser;
+module.exports.getUserById = getUserById;
 
 module.exports.addRepoToProject = addRepoToProject;
 module.exports.getRepoById = getRepoById;
